+++ coolercontrold/src/repositories/gpu/amd.rs
@@ -159,7 +159,7 @@ impl GpuAMD {
                     model,
                     u_id,
                     channels,
-                    block_dev_path: None,
+                    ..Default::default()
                 },
                 device_path,
                 fan_curve_info,
@@ -632,7 +632,7 @@ impl GpuAMD {
                     channel.hwmon_type == HwmonChannelType::Fan && channel.name == channel_name
                 })
                 .with_context(|| format!("Searching for channel name: {channel_name}"))?;
-            fans::set_pwm_enable_to_default(&amd_hwmon_info.hwmon.path, channel_info).await
+            fans::set_pwm_enable_to_default_or_auto(&amd_hwmon_info.hwmon.path, channel_info).await
         }
     }