+++ coolercontrold/src/repositories/hwmon/temps.rs
@@ -18,7 +18,7 @@
 
 use std::io::{Error, ErrorKind};
 use std::ops::Not;
-use std::path::{Path, PathBuf};
+use std::path::Path;
 
 use crate::cc_fs;
 use crate::device::TempStatus;
@@ -34,7 +34,7 @@ const TEMP_SANITY_MIN: f64 = 0.0;
 const TEMP_SANITY_MAX: f64 = 140.0;
 
 /// Initialize all applicable temp sensors
-pub async fn init_temps(base_path: &PathBuf, device_name: &str) -> Result<Vec<HwmonChannelInfo>> {
+pub async fn init_temps(base_path: &Path, device_name: &str) -> Result<Vec<HwmonChannelInfo>> {
     if temps_used_by_another_repo(device_name) {
         return Ok(vec![]);
     }
@@ -275,7 +275,6 @@ mod tests {
             assert_eq!(temps[0].hwmon_type, HwmonChannelType::Temp);
             assert_eq!(temps[0].name, "temp1");
             assert_eq!(temps[0].label, Some("Temp 1".to_string()));
-            assert!(!temps[0].pwm_mode_supported);
             assert_eq!(temps[0].pwm_enable_default, None);
             assert_eq!(temps[0].number, 1);
         });